requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
filelock==3.13.4
fsspec==2024.3.1
huggingface-hub==0.22.2
idna==3.7
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pytorch-crf==0.7.2
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
safetensors==0.4.3
sympy==1.12
tokenizers==0.19.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.2
transformers==4.40.0
typing_extensions==4.11.0
urllib3==2.2.1
